# Optimize 15-34% by hoisting indexing to induce loop unrolling